requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
anthropic==0.5.0
langchain==0.0.335
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.8.0
scikit-learn==1.3.0
nltk==3.8.1
PyPDF2==3.0.1
cohere==4.37
google-generativeai==0.3.1

# Note: EmergenceAI and Groq don't have official Python clients at this time
# We're using requests directly for their APIs
